pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mod-gpt"
version = "0.1.0"
description = "LLM-powered Discord moderation experiment with agent routing and function-calling actions."
authors = [{ name = "Luke Oliff" }]
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "asyncpg>=0.29.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "B", "I"]
ignore = ["E203", "E501"]

[tool.uv]
dev-dependencies = [
    "black>=23.10.0",
    "ruff>=0.1.5",
]

[tool.uv.scripts]
bot = "python main.py"
